catering-backend/requirements.txt
Django==1.6.5
South==1.0
argparse==1.2.1
django-cors-headers==0.12
djangorestframework==2.3.14
wsgiref==0.1.2
